Use jsdom instead of directly contextify.
After some experimentation I've reached the conclusion that running
AngularJS outside of the context of a DOM is a fool's errand, so instead
we'll provide it with just enough DOM to keep it happy.

This removes the requirement for a patched version of AngularJS, and also
allows the $compile service to work out of the box because it is able
to create elements using our jsdom document as the owner.

The downside here is that a context object has considerably more startup
and memory overhead, but being able to use $compile is a key part of
most potential applications of this module, so we accept that overhead.