Provide access to the global 'require' object too.
Using this is necessary in order to successfully load code that uses
RequireJS, since that requires a custom implementation of require.load
that knows how to load modules in the node world. Later it would be nice
to provide such an implementation that uses runFile, but for the moment
the caller needs to provide that.