Put a dummy script element into the internal DOM.
This really shouldn't be necessary, but some apps unconditionally insert
analytics scripts into the DOM by searching for an existing script
and inserting a new script after it. Having this script element here makes
that insert work, although the analytics script won't actually run because
we only execute scripts that are explicitly provided via runFiles.

The main goal here is just to allow existing apps to run with few
modifications... but an app optimized for use in node would have an
entry point that only loads what's needed in the server context, and skips
the stuff that only makes sense in a browser.