Update nodejs versions used for Travis-CI tests
Node 4 and 5 are now obsolete, and so there's no need to run tests against them.

Versions 6, 8 and 10 are still within support as of this commit, with 6 in LTS maintenance and 8 and 10 under active development. Node 9 will be obsolete in a few weeks, so we'll just skip adding it in the first place.